requirements.txt
pandas 
notebook
numpy
scikit-learn
matplotlib
mlflow==2.2.2
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
# -e .